Add the two CubeMX projects
As a start, I include the Practise Enterprise files for the STM32F7 Series. "Practise Enterprise 2" was intended for using the STM32F7508. But since there is not enough FLASH memory on the chip and the built-in QuadSPI chip also had some issues with time-critical events, I needed to move onto the STM32F746G. This chip has a lot more FLASH memory whereby my project worked from the first time again. The "PE2_Clone" is the project used for the F746.