Added first try of this year
This is the first try of this year. I removed all features that I don't need anymore. I added some things to optimize it for scanning (and not for sending anymore). I also included a selfmade website where I can calculate the code that needs to be sent to the I/O expander for different type of events.